son hazırlıklar yapıldı.

+++ src/main/java/model/Ders.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Ders {
-    private int id;
+    private Long id;
     private String ad;
     private int ogretmen_id;
 
@@ -9,14 +9,14 @@ public class Ders {
     public Ders() {}
 
 
-    public Ders(int id, String ad, int  ogretmen_id) {
+    public Ders(Long id, String ad, int  ogretmen_id) {
         this.id = id;
         this.ad = ad;
         this.ogretmen_id = ogretmen_id;
     }
 
-    public  int getId() {return id;}
-    public void setId(int id) {this.id = id;}
+    public  Long getId() {return id;}
+    public void setId(Long id) {this.id = id;}
 
     public String getAd() {return ad;}
     public void setAd(String ad) {this.ad = ad;}
